Remove usage of `LazilyInitializedFileCollection`
Duplicating the code here removes the usage of the internal API. We also
don't need `TaskDependencyResolveContext` anymore.
There is a slight change: if the task is being requested (part of the
task graph), then the `FileCollection` is created when querying the
dependencies of the task when building the task graph instead of just
before running the task.